Added CoseAlgorithmManager to AssertionResponseValidator
Had a hard time trying to know why following this PublicKeyCredentialRequest Response Verification example literally wasn't working in my tests.

Well, if $coseAlgorithmManager parameter is not present in "new AuthenticatorAssertionResponseValidator", then it's obviously null and, when that happens, the default algorithm is SHA256. Probably not the same with the one that your authenticator is using.